==> explore/Seurat_workflow/05_Population_cluster_check.Rmd <==
---
title: "05_Population_cluster_check"
output: html_document
---

```{r setup, include=FALSE}
library(Seurat)
library(here)
library(tidyverse)
library(cowplot)
library(knitr)

library(BiocManager)
library(sctransform)
library(glmGamPoi)
library(Matrix)
library(future)
library(future.apply)
```

```{r}
plan('multiprocess', workers = 20)
options(future.globals.maxSize = 2e+05 * 1024^2)
```

```{r}
refquery <- readRDS(here('explore/RDS/05_Integrate_df_annotation.RDS'))
head(refquery)
```

```{r}
head(refquery)
```

```{r}
unique(refquery@meta.data$cell_type_main)
```

- Epithelial

- Immune

- Endothelial

- Mesenchymal


# Epithelial cells
```{r}
#refquery <- readRDS(here('explore/RDS/05_Integrate_df_annotation.RDS'))

refquery <- subset(refquery, cell_type_main == 'Epithelial cells')
refquery <- ScaleData(object = refquery)
refquery <- RunPCA(object = refquery)
refquery <- FindNeighbors(refquery, dims = 1:20)
refquery <- FindClusters(refquery, resolution = 0.6)
refquery <- RunUMAP(object = refquery, dims = 1:20)

unique(refquery@meta.data$integrated_snn_res.0.6)
unique(refquery@meta.data$integrated_snn_res.0.8)
unique(refquery@meta.data$seurat_clusters)

rm(refquery)
```


# Immune cells
```{r}
refquery <- readRDS(here('explore/RDS/05_Integrate_df_annotation.RDS'))

refquery <- subset(refquery, cell_type_main == 'Immune cells')
refquery <- ScaleData(object = refquery)
refquery <- RunPCA(object = refquery)
refquery <- FindNeighbors(refquery, dims = 1:20)
refquery <- FindClusters(refquery, resolution = 0.6)
refquery <- RunUMAP(object = refquery, dims = 1:20)

unique(refquery@meta.data$integrated_snn_res.0.6)
unique(refquery@meta.data$integrated_snn_res.0.8)
unique(refquery@meta.data$seurat_clusters)

rm(refquery)
```


# Endothelial cells 
```{r}
refquery <- readRDS(here('explore/RDS/05_Integrate_df_annotation.RDS'))

refquery <- subset(refquery, cell_type_main == 'Endothelial cells')
refquery <- ScaleData(object = refquery)
refquery <- RunPCA(object = refquery)
refquery <- FindNeighbors(refquery, dims = 1:20)
refquery <- FindClusters(refquery, resolution = 0.6)
refquery <- RunUMAP(object = refquery, dims = 1:20)

unique(refquery@meta.data$integrated_snn_res.0.6)
unique(refquery@meta.data$integrated_snn_res.0.8)
unique(refquery@meta.data$seurat_clusters)

rm(refquery)
```


# Mesenchymal cells
```{r}
refquery <- readRDS(here('explore/RDS/05_Integrate_df_annotation.RDS'))

refquery <- subset(refquery, cell_type_main == 'Mesenchymal cells')
refquery <- ScaleData(object = refquery)
refquery <- RunPCA(object = refquery)
refquery <- FindNeighbors(refquery, dims = 1:20)
refquery <- FindClusters(refquery, resolution = 0.6)
refquery <- RunUMAP(object = refquery, dims = 1:20)

unique(refquery@meta.data$integrated_snn_res.0.6)
unique(refquery@meta.data$integrated_snn_res.0.8)
unique(refquery@meta.data$seurat_clusters)

rm(refquery)
```
